Move adapters and viewholders to own package
