--- app/src/main/java/be/ugent/zeus/hydra/ui/common/recyclerview/DataViewHolderFactory.java
@@ -1,7 +1,0 @@
-package be.ugent.zeus.hydra.ui.common.recyclerview;
-
-import android.view.View;
-
-public interface DataViewHolderFactory<E> {
-    DataViewHolder<E> newInstance(View itemView);
-}
