+++ app/src/main/java/be/ugent/zeus/hydra/ui/common/recyclerview/adapters/DiffAdapter.java
@@ -1,9 +1,12 @@
-package be.ugent.zeus.hydra.ui.common.recyclerview;
+package be.ugent.zeus.hydra.ui.common.recyclerview.adapters;
 
 import android.os.Handler;
 import android.os.Looper;
 import android.support.v7.util.DiffUtil;
 import android.support.v7.widget.RecyclerView;
+
+import be.ugent.zeus.hydra.ui.common.recyclerview.EqualsCallback;
+import be.ugent.zeus.hydra.ui.common.recyclerview.adapters.Adapter;
 import java8.util.function.BiFunction;
 
 import java.util.ArrayList;
