+++ app/src/main/java/be/ugent/zeus/hydra/ui/common/recyclerview/viewholders/DataViewHolder.java
@@ -1,7 +1,9 @@
-package be.ugent.zeus.hydra.ui.common.recyclerview;
+package be.ugent.zeus.hydra.ui.common.recyclerview.viewholders;
 
 import android.support.v7.widget.RecyclerView;
 import android.view.View;
+
+import be.ugent.zeus.hydra.ui.common.recyclerview.adapters.Adapter;
 
 /**
  * A generic view holder, intended as root of all other view holders.
