+++ app/src/main/java/be/ugent/zeus/hydra/ui/common/recyclerview/viewholders/DataViewHolderFactory.java
@@ -1,0 +1,9 @@
+package be.ugent.zeus.hydra.ui.common.recyclerview.viewholders;
+
+import android.view.View;
+
+import be.ugent.zeus.hydra.ui.common.recyclerview.viewholders.DataViewHolder;
+
+public interface DataViewHolderFactory<E> {
+    DataViewHolder<E> newInstance(View itemView);
+}
