+++ app/src/main/java/be/ugent/zeus/hydra/ui/common/recyclerview/viewholders/DateHeaderViewHolder.java
@@ -1,4 +1,4 @@
-package be.ugent.zeus.hydra.ui.common.recyclerview;
+package be.ugent.zeus.hydra.ui.common.recyclerview.viewholders;
 
 import android.support.v7.widget.RecyclerView;
 import android.view.View;
