+++ app/src/main/java/be/ugent/zeus/hydra/ui/main/EventAdapter.java
@@ -4,8 +4,8 @@ import android.view.ViewGroup;
 
 import be.ugent.zeus.hydra.R;
 import be.ugent.zeus.hydra.data.models.association.Event;
-import be.ugent.zeus.hydra.ui.common.recyclerview.DateHeaderViewHolder;
-import be.ugent.zeus.hydra.ui.common.recyclerview.ItemAdapter;
+import be.ugent.zeus.hydra.ui.common.recyclerview.viewholders.DateHeaderViewHolder;
+import be.ugent.zeus.hydra.ui.common.recyclerview.adapters.ItemAdapter;
 import be.ugent.zeus.hydra.ui.common.ViewUtils;
 import com.timehop.stickyheadersrecyclerview.StickyRecyclerHeadersAdapter;
