+++ app/src/main/java/be/ugent/zeus/hydra/ui/main/InfoViewHolder.java
@@ -8,7 +8,7 @@ import android.widget.TextView;
 
 import be.ugent.zeus.hydra.R;
 import be.ugent.zeus.hydra.data.models.info.InfoItem;
-import be.ugent.zeus.hydra.ui.common.recyclerview.DataViewHolder;
+import be.ugent.zeus.hydra.ui.common.recyclerview.viewholders.DataViewHolder;
 import be.ugent.zeus.hydra.ui.common.ViewUtils;
 
 import static be.ugent.zeus.hydra.ui.common.ViewUtils.$;
