+++ app/src/main/java/be/ugent/zeus/hydra/ui/main/MinervaCourseViewHolder.java
@@ -7,7 +7,7 @@ import android.widget.TextView;
 import be.ugent.zeus.hydra.R;
 import be.ugent.zeus.hydra.ui.minerva.overview.CourseActivity;
 import be.ugent.zeus.hydra.data.models.minerva.Course;
-import be.ugent.zeus.hydra.ui.common.recyclerview.DataViewHolder;
+import be.ugent.zeus.hydra.ui.common.recyclerview.viewholders.DataViewHolder;
 import be.ugent.zeus.hydra.ui.common.html.Utils;
 
 import static be.ugent.zeus.hydra.ui.common.ViewUtils.$;
