+++ app/src/main/java/be/ugent/zeus/hydra/ui/main/homefeed/content/HideableViewHolder.java
@@ -6,7 +6,7 @@ import android.widget.TextView;
 import be.ugent.zeus.hydra.BuildConfig;
 import be.ugent.zeus.hydra.R;
 import be.ugent.zeus.hydra.ui.main.homefeed.HomeFeedAdapter;
-import be.ugent.zeus.hydra.ui.common.recyclerview.DataViewHolder;
+import be.ugent.zeus.hydra.ui.common.recyclerview.viewholders.DataViewHolder;
 import be.ugent.zeus.hydra.ui.common.widgets.NowToolbar;
 
 import static be.ugent.zeus.hydra.ui.common.ViewUtils.$;
