+++ app/src/main/java/be/ugent/zeus/hydra/ui/minerva/overview/AgendaAdapter.java
@@ -3,8 +3,8 @@ package be.ugent.zeus.hydra.ui.minerva.overview;
 import android.view.ViewGroup;
 import be.ugent.zeus.hydra.R;
 import be.ugent.zeus.hydra.data.models.minerva.AgendaItem;
-import be.ugent.zeus.hydra.ui.common.recyclerview.EmptyItemAdapter;
-import be.ugent.zeus.hydra.ui.common.recyclerview.DateHeaderViewHolder;
+import be.ugent.zeus.hydra.ui.common.recyclerview.adapters.EmptyItemAdapter;
+import be.ugent.zeus.hydra.ui.common.recyclerview.viewholders.DateHeaderViewHolder;
 import be.ugent.zeus.hydra.ui.common.ViewUtils;
 import com.timehop.stickyheadersrecyclerview.StickyRecyclerHeadersAdapter;
 import org.threeten.bp.format.DateTimeFormatter;
