+++ app/src/main/java/be/ugent/zeus/hydra/ui/minerva/overview/AnnouncementAdapter.java
@@ -3,7 +3,7 @@ package be.ugent.zeus.hydra.ui.minerva.overview;
 import android.view.ViewGroup;
 import be.ugent.zeus.hydra.R;
 import be.ugent.zeus.hydra.data.models.minerva.Announcement;
-import be.ugent.zeus.hydra.ui.common.recyclerview.EmptyItemAdapter;
+import be.ugent.zeus.hydra.ui.common.recyclerview.adapters.EmptyItemAdapter;
 import be.ugent.zeus.hydra.ui.common.ViewUtils;
 
 /**
