+++ app/src/main/java/be/ugent/zeus/hydra/ui/sko/overview/LineupViewHolder.java
@@ -13,7 +13,7 @@ import android.widget.TextView;
 import be.ugent.zeus.hydra.R;
 import be.ugent.zeus.hydra.ui.sko.ArtistActivity;
 import be.ugent.zeus.hydra.data.models.sko.Artist;
-import be.ugent.zeus.hydra.ui.common.recyclerview.DataViewHolder;
+import be.ugent.zeus.hydra.ui.common.recyclerview.viewholders.DataViewHolder;
 import com.squareup.picasso.Picasso;
 import org.threeten.bp.LocalDateTime;
 import org.threeten.bp.format.DateTimeFormatter;
